Fix postLikes and getLikes function
