+++ src/modules/likes/getLikes.js
@@ -1,0 +1,13 @@
+const getTotalLikes = async () => {
+  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/yOdv0JXV8ssnN9f9thVW/likes', {
+    method: 'GET',
+    headers: {
+      Accept: 'application/json',
+    },
+  });
+  const data = await response.json();
+  // const results = data.find((item) => +item.item_id === +id);
+  return data;
+};
+
+export default getTotalLikes;
