+++ src/modules/likes/postLikes.js
@@ -1,0 +1,16 @@
+// POST ACTION
+const addLike = async (id) => {
+  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/yOdv0JXV8ssnN9f9thVW/likes', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      item_id: id,
+    }),
+  });
+  const jsonResponse = await response.json();
+  return jsonResponse;
+};
+
+export default addLike;
